Created application to automatically upload files to YouTube
Tracks a given directory, waiting for the creation of a new file - such as one created through beginning an OBS recording

It then periodically checks for the last time the file was written to. If it does not detect a change in the file's state, we determine that the file is ready to upload.

Calls YoutubeAPI using Env Vars for Authorization